MCP/requirements.txt
mcp[cli]
python-dotenv
google-generativeai
langchain-mistralai
mistralai
ipykernel
httpx 
uv